Retrofit the Yk JIT into the Lua interpreter.
Obviously nothing really works yet, but this is a good starting point.

The JIT can be conditionally compiled in if `YK_DIR` is set (e.g. `make
YK_DIR=~/yk`). Sadly I had to make a portable Makefile into a GNU
Makefile for that to work though. I don't think it matters too much.

Currently we only make JIT locations for regular `for` loops.

The test suite currently seg faults when using the JIT, so that's
commented in CI for now.

This simple program makes it to stop-gapping and then fails a
stackmap-related assertion:

```
for i = 0, 100 do
    print("hello!")
end
```

It's a start.