Only construct one MT instance per Lua interpreter instance.
Previously we constructed a new MT instance for every Lua function call: this
commit moves that to what appears to be the global "initialise/deinitialise a
Lua interpreter" functions. I'm not entirely sure if I've got the
`yk_mt_drop` in exactly the right part of the `close` function, but it's
certainly nearer to the right place than before.