Optimise the YkLocation buffer.
Location array was reallocated for every instructions, which was very ineffective
since we only need to reallocate the buffer either when yk location is set to NULL
or when the number of bytecode instructions is greater than number of location
in the yk location array. This was causing around 1000x slowdown in benchmark
test such as verybig.lua. The fix improves verybig.lua by 300x and
brings performance from 198sec to 0.650sec when JIT is off.