Promote immediates.
This is a conservative use of `yk_promote` operating on Lua opcodes
which themselves use immediates (e.g. OP_ADDI). It does have a small
(just over 1% improvement) effect on big_loop.lua, but the main point of
this is to show that we can start to move an interpreter in a
yk-specific direction.