Use the `pcRel` macro to compute the current pc.
The line of code in this commit was off-by-one, though in a way that
didn't undermine the correctness of yk per se, since we were always
off-by-one. However, it would have meant that we wrote to the word
immediately past the end of an allocation, which is UB.

The cause of this is the `vmfetch` macro, executed before the line of
code changed by this commit which increments `pc`: thus by the time we
execute the line of code in this commit, `pc` no longer points to the
actual `pc` we wanted, but one after it. The `pcRel` macro does the
necessary magic (i.e. subtracting 1) to resolve this.