Use the yklua-tests repo.
Before we had a submodule for the openresty tests and a duplicated
run.sh. This removes the submodule and the duplicated runner and
switches instead to using our yklua-tests repo.

While we are here, also test unserialised compilation.

Note: serialised compilation still takes a long time to test on account
of tak.lua being so very slow.