Properly assign YkLocations when undumping a function.
Before, we'd make a location for *every* pc in an unmdumped function.
This could lead to us tracing too much.

This fixes this by using the same heuristics that we use for normal
"parsed" functions to decide where loops start, and thus where to put
yk locations.

(Ideally we'd dump and undump the locations, thus retaining JIT state,
but that could be tricky. One for another day)

In terms of performance, most benchmarks either don't change, or speed
up a little bit (in the order of 5%).