Add `YkLocation`s to the start of functions.
The main effect of this to enable recursive functions to be traced as
loops. This has a significant effect on programs where these are the
main kinds of loops (e.g. `towers` speeds up by ~8.5x). It does have a
small slowdown for some other benchmarks (at worst a 1.5x overhead for
Deltablue), but the wins outweigh the losses -- and, probably, we can
claw back some or all of those losses by refining this approach (both
within yklua and yk itself) in the future.