Add `os.monotonic` to expose a monotonic timer.
`os.clock` uses libc's `clock` which measure "user + sys" time which is
very rarely useful. This patch adds a new timer `os.monotonic` which
returns a monotonic timer. On Linux, none of the monotonic timers is
documented to do exactly what we would like: of those available,
`MONOTONIC` seems the option least likely to produce misleading results,
so we pick that.