Hoist proto_version load.
Before this commit, we loaded a closure's `proto_version` on every
iteration of the loop, which leads to a guard in every opcode in a
trace.

This commit hoists that load out to the only point where that value can
change -- the `startfunc`/`return` label(s). These are where a new
closure is dealt with (e.g. as a result of `OP_CALL`). Although we still
have to call `yk_promote` for every opcode, the trace optimiser has a
much easier job, because this tends to stay in a variable -- in general
it can now remove 1 guard for all but the first opcode. Across several
benchmarks I've looked at, this leads to a 2-6% increase.