Dynamically detect recursion.
Previously we added a `YkLocation` to the start of every Lua function, which
meant that we created lots of (mostly) pointless traces.

This commit changes things so that only when we dynamically detect that
a Lua function is recursive do we add a `YkLocation` to it. This largely
prevents us tracing pointless functions.

This fix is done in constant time using a trick that took me a long
while to realise: we only need to detect that we're in one level of
recursion and we only need to do so the first time that happens. A
single boolean suffices for this (previously I'd tried being much more
precise, but (a) that required examining every call on the stack (b) it
turns out not to be necessary): it's always accurate for the first
recursive call, and only becomes inaccurate after that point (at which
point we don't need it anyway).