Split `setnodevector`.
The common case, by far, is `size == 0` so this function was being
outlined mostly for no good reason. This commit splits it into two: the
common case can be inlined by yk, with the uncommon case leading to a
`call`.

This leads to a small, but just about measurable speed up, on some
benchmarks (including ~3% on Richards).