Mark all GC functions as `yk_outline`.
There are few, if any, GC functions that we want to trace directly but
at the moment, SWT inserts tracing blocks which (a) slows these
functions down in all cases (b) particularly slows them down if tracing
is active.

At first I started going through these individually, seeing which end up
being outlined before realising that really what I want to do is say
"all of the functions in this file should be outlined". I don't believe
there's a way to attach such an attribute to a file in LLVM -- or, at
least, we don't have the right hooks for it in ykllvm right now -- so
this commit simply marks every function in this file as `yk_outline`.
It wouldn't surprise me if, in the future, we come back and unmark one
or two functions, but I doubt we'll need or want to do that soon.

Across our benchmark suite this leads to a sizeable performance
increase, with many benchmarks improving by around 8-9%.